Added possibility for parallel ordered merge
