+++ src/main/java/hu/akarnokd/rxjava2/parallel/ParallelTransformers.java
@@ -16,7 +16,14 @@
 
 package hu.akarnokd.rxjava2.parallel;
 
+import java.util.Comparator;
+
+import io.reactivex.Flowable;
+import io.reactivex.internal.functions.Functions;
+import io.reactivex.internal.functions.ObjectHelper;
+import io.reactivex.parallel.ParallelFlowable;
 import io.reactivex.parallel.ParallelTransformer;
+import io.reactivex.plugins.RxJavaPlugins;
 
 /**
  * Transformers for RxJava 2 ParallelFlowable sequences.
@@ -26,6 +33,96 @@ public final class ParallelTransformers {
 
     private ParallelTransformers() {
         throw new IllegalStateException("No instances!");
+    }
+    
+    /**
+     * Merges the source ParallelFlowable rails in an ordered fashion picking the smallest of the available value from
+     * them (determined by their natural order).
+     * @param <T> the value type of all sources
+     * @param source the source ParallelFlowable
+     * @return the new Flowable instance
+     */
+    public static <T extends Comparable<? super T>> Flowable<T> orderedMerge(ParallelFlowable<T> source) {
+        return orderedMerge(Functions.naturalOrder(), false, Flowable.bufferSize(), source);
+    }
+
+
+    /**
+     * Merges the source ParallelFlowable rails in an ordered fashion picking the smallest of the available value from
+     * them (determined by their natural order) and allows delaying any error they may signal.
+     * @param <T> the value type of all sources
+     * @param source the source ParallelFlowable
+     * @param delayErrors if true, source errors are delayed until all sources terminate in some way
+     * @return the new Flowable instance
+     */
+    public static <T extends Comparable<? super T>> Flowable<T> orderedMerge(boolean delayErrors, ParallelFlowable<T> source) {
+        return orderedMerge(Functions.naturalOrder(), delayErrors, Flowable.bufferSize(), source);
+    }
+
+
+    /**
+     * Merges the source ParallelFlowable rails in an ordered fashion picking the smallest of the available value from
+     * them (determined by their natural order), allows delaying any error they may signal and sets the prefetch
+     * amount when requesting from these Publishers.
+     * @param <T> the value type of all sources
+     * @param source the source ParallelFlowable
+     * @param delayErrors if true, source errors are delayed until all sources terminate in some way
+     * @param prefetch the number of items to prefetch from the sources
+     * @return the new Flowable instance
+     */
+    public static <T extends Comparable<? super T>> Flowable<T> orderedMerge(boolean delayErrors, int prefetch, ParallelFlowable<T> source) {
+        return orderedMerge(Functions.naturalOrder(), delayErrors, prefetch, source);
+    }
+
+
+    /**
+     * Merges the source ParallelFlowable rails in an ordered fashion picking the smallest of the available value from
+     * them (determined by the Comparator).
+     * @param <T> the value type of all sources
+     * @param source the source ParallelFlowable
+     * @param comparator the comparator to use for comparing items;
+     *                   it is called with the last known smallest in its first argument
+     * @return the new Flowable instance
+     */
+    public static <T> Flowable<T> orderedMerge(Comparator<? super T> comparator, ParallelFlowable<T> source) {
+        return orderedMerge(comparator, false, Flowable.bufferSize(), source);
+    }
+
+
+    /**
+     * Merges the source ParallelFlowable rails in an ordered fashion picking the smallest of the available value from
+     * them (determined by the Comparator) and allows delaying any error they may signal.
+     * @param <T> the value type of all sources
+     * @param source the source ParallelFlowable
+     * @param comparator the comparator to use for comparing items;
+     *                   it is called with the last known smallest in its first argument
+     * @param delayErrors if true, source errors are delayed until all sources terminate in some way
+     * @return the new Flowable instance
+     * 
+     */
+    public static <T> Flowable<T> orderedMerge(Comparator<? super T> comparator, boolean delayErrors, ParallelFlowable<T> source) {
+        return orderedMerge(comparator, delayErrors, Flowable.bufferSize(), source);
+    }
+
+
+    /**
+     * Merges the source ParallelFlowable rails in an ordered fashion picking the smallest of the available value from
+     * them (determined by the Comparator), allows delaying any error they may signal and sets the prefetch
+     * amount when requesting from these Publishers.
+     * @param <T> the value type of all sources
+     * @param source the source ParallelFlowable
+     * @param comparator the comparator to use for comparing items;
+     *                   it is called with the last known smallest in its first argument
+     * @param delayErrors if true, source errors are delayed until all sources terminate in some way
+     * @param prefetch the number of items to prefetch from the sources
+     * @return the new Flowable instance
+     * 
+     */
+    public static <T> Flowable<T> orderedMerge(Comparator<? super T> comparator, boolean delayErrors, int prefetch, ParallelFlowable<T> source) {
+        ObjectHelper.requireNonNull(comparator, "comparator is null");
+        ObjectHelper.requireNonNull(source, "sources is null");
+        ObjectHelper.verifyPositive(prefetch, "prefetch");
+        return RxJavaPlugins.onAssembly(new ParallelOrderedMerge<T>(source, comparator, delayErrors, prefetch));
     }
 
     /**
